requirements.txt
Flask==2.2.2
gunicorn
requests
python-dotenv